[LB-464] feat : 트레이싱 설정
