+++ src/main/java/com/example/tutor_bot/config/KafkaConfig.java
@@ -1,0 +1,9 @@
+package com.example.tutor_bot.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@Configuration
+@EnableKafka
+public class KafkaConfig {
+}
